♻️ Opprett `GjennopprettAvklarteDagerService.kt` og del opp logikk til flere filer
